Allow configurable flag sources
This change allows us to define `FLAG_SOURCES` as a dotted list of Python paths to classes that provide a `get_flags()` method that returns a dictionary formatted like the `FLAGS` setting.

This opens up potential for supporting sources of flags and conditions that are not just the settings files or the database, and lets someone disable either database or settings flags as a thing entirely.

This is a breaking change for `FLAG_SOURCES`, which previously supported only module paths that defined uppercase, boolean variables only. This use of `FLAG_SOURCES` has never been documented as a supported feature.