Move flag wrapping into FlaggedViewMixin.as_view
This moves the wrapping in a feature flag (and thus the point at which bad configuration will cause errors) into `as_view()`. This will allow view/fallback introspection, as well as ensure that errors happen when the view is created, rather than when requests are dispatched to it (which is the same as using the decorators on non-class based views).