Use clearer language for booleans with required conditions
This change will say 'MYFLAG is disabled for all requests, even when all required conditions are met.' when there are required conditions along with a non-required boolean condition and no other non-required conditions.

This is a change from 'MYFLAG is disabled when all required conditions are met.' which is somewhat confusing.

This change also refactors the string creation somewhat to reduce the duplication of strings. Hopefully the way I'm slicing makes eventual translation doable... but I'm not entirely sure that'll be the case.