Remove "optional" from state string language
"Optional" does not seem like the right word to use in retrospect. The difference between required and non-required conditions is the difference between a boolean AND and a boolean OR. So, no condition is "optional" — at least one of them has to be met.

This change removes the word "optional" from the state string when there are no required conditions, and replaces it with "non-required" when there are.