Add support for Django 3.0
This change adds support for Django 3.0. It requires updating the Django Rest Framework supported version range and dealing with the minor API changes that it brings with it.

This change also drops testing for non-LTS versions of Django 2.0. Those versions should be supported, but if you're using Django 2.x you should be on 2.2.