Deprecate middleware, instead lazily cache flags
Currently use of FlagConditionsMiddleware means that flags are retrieved
on every single Django request, even those that don't need to check
flags. If developers don't use this middleware, they instead have to
retrieve flags each time they are checked.

It would be nice if instead flags were lazily cached, so that the first
time they are retrieved, they get cached, and all subsequent checks use
the cache. For requests without the need to check flags, however, the
unnecessary call to the database would be skipped, and they'd never be
retrieved.

This commit makes that change, incorporating the logic into the existing
get_flags method.