Introduce validators for conditions
This change supports adding a `validator` callable when registering a condition. This validator is used by the `FlagStateForm` to ensure that values entered for conditions are valid for that condition, and to provide feedback if not.

To support that functionality, this change also refactors the conditions code, breaking the registry and the built-in conditions into their own modules, and adding a new module for validators. The API remains the same.