Support specifying metadata in SettingsFlagsSource
This change supports specifying metadata in the SettingsFlagsSource. To support that, this change will also deprecate the previous behavior of flag sources' `get_flags` method, which was expected to return a dictionary, and instead expect a three-tuple of (flag, conditions, metadata).

A warning will be raised that `get_flags` returning a dict is deprecated and support for doing so will be removed in 6.0.