Remove six from isort standard libraries
isort seems to have switched to `frozenset` for `known_standard_library`, which broke the way we're overriding it here to add `six`. This caused only `six` to be recognized as a standard library, and everything else as third-party. This is an explicit behavior change (see https://github.com/timothycrosley/isort/issues/1238).

We no longer use `six` so we can remove this setting