Updates to support isort 5.0 (#70)
Removed `not_skip` and `recursive` from `tox.ini` to fix
TypeError: __init__() got an unexpected keyword argument 'not_skip'

* recursive
Prior to version 5.0.0, isort wouldn't automatically traverse directories. The --recursive option was necessary to tell it to do so. In 5.0.0 directories are automatically traversed for all Python files, and as such this option is no longer necessary and should simply be removed.

* not_skip
In an earlier version isort had a default skip of __init__.py. To get around that many projects wanted a way to not skip __init__.py or any other files that were automatically skipped in the future by isort. isort no longer has any default skips, so if the value here is __init__.py you can simply remove the setting.

https://timothycrosley.github.io/isort/docs/upgrade_guides/5.0.0/

* Add python_requires to setup.py

* Remove six from isort standard libraries

isort seems to have switched to `frozenset` for `known_standard_library`, which broke the way we're overriding it here to add `six`. This caused only `six` to be recognized as a standard library, and everything else as third-party. This is an explicit behavior change (see https://github.com/timothycrosley/isort/issues/1238).

We no longer use `six` so we can remove this setting

* Remove known_future_library line in isort config in tox.ini