Fix an issue with `flagged_path` and includes
This change fixes an issue with `flagged_path` where `includes()` would not be correctly included, and cannot be resolved using their names.

The issue was simply with our passing of `is_endpoint` to the pattern class. In the case of `path()` URLs, the pattern class is `RoutePattern`, and it doesn’t double-check `is_endpoint` the way `RegexPattern` does. I have some more details in #100.

This is a simple fix to change that incorrect `True` to `False`.